requirements.txt
playwright==1.44.0
httpx==0.27.0
apscheduler==3.10.4
gspread==6.1.2
google-auth==2.30.0
python-dotenv==1.0.1
notion-client==2.2.1
